+++ src/components/sykmeldinger/Pengestopp.tsx
@@ -8,10 +8,10 @@ import { connect } from 'react-redux';
 import {
     Status,
     StatusEndring,
+    Sykmelding,
 } from '../../types/FlaggPerson';
 import { initialState as FlaggPerson } from '../../reducers/flaggperson';
-import { senesteTom } from "../../utils/periodeUtils";
-import { gamleSMStatuser } from "../../utils/sykmeldinger/sykmeldingstatuser";
+import { allStoppAutomatikkStatusEndringer, arbeidsgivereWithStoppAutomatikkStatus, sykmeldingerToArbeidsgiver, unikeArbeidsgivereMedSykmeldingSiste3Maneder, uniqueArbeidsgivere } from '../../utils/pengestoppUtils';
 
 const texts = {
     stansSykepenger: 'Stans sykepenger',
@@ -21,79 +21,13 @@ const Wrapper = styled.div`
     margin: 1rem 0;
 `;
 
-export interface Arbeidsgiver {
-    navn: string,
-    orgnummer: string,
-}
-
-export interface Periode {
-    tom: Date,
-}
-
-export interface Sykmelding {
-    arbeidsgiver: string,
-    orgnummer: string,
-    pasient: {
-        fnr: string
-    },
-    mulighetForArbeid: {
-        perioder: Periode[]
-    }
-    status: String,
-}
-
-export interface IPengestoppProps {
+interface IPengestoppProps {
     brukernavn: String,
-    sykmeldinger: Array<Sykmelding>,
+    sykmeldinger: Sykmelding[],
     flaggperson: typeof FlaggPerson,
     fnr: string,
 }
 
-const sykmeldingerToArbeidsgiver = (sykmeldinger: Array<Sykmelding>) => {
-    return sykmeldinger.map((sykmelding) => {
-        return {
-            navn: sykmelding.arbeidsgiver,
-            orgnummer: sykmelding.orgnummer,
-        };
-    });
-};
-
-const uniqueArbeidsgivere = (arbeidsgivere: Array<Arbeidsgiver>) => {
-    return arbeidsgivere.filter((arbeidsgiver, index, self) => {
-        return self.findIndex((arbeidsgiver2) => {
-            return arbeidsgiver.orgnummer === arbeidsgiver2.orgnummer;
-        }) === index;
-    });
-};
-
-const allStoppAutomatikkStatusEndringer = (statusEndringer: Array<StatusEndring>) => {
-    return statusEndringer.filter((statusEndring) => {
-        return statusEndring.status === Status.STOPP_AUTOMATIKK;
-    });
-};
-
-const arbeidsgivereWithStoppAutomatikkStatus = (arbeidsgivere: Array<Arbeidsgiver>, statusEndringerWithStoppAutomatikk: Array<StatusEndring>) => {
-    return arbeidsgivere.filter((arbeidsgiver) => {
-        return statusEndringerWithStoppAutomatikk.find((statusEndring) => {
-            return statusEndring.virksomhetNr.value === arbeidsgiver.orgnummer;
-        });
-    });
-};
-
-const aktiveSykmeldingerFraSiste3Maneder = (sykmeldinger: Sykmelding[]) => {
-    const threeMonthsAgo = new Date(Date.now() - 1000 * 60 * 60 * 24 * 90);
-    return sykmeldinger.filter((sykmelding) => {
-        return senesteTom(sykmelding.mulighetForArbeid.perioder) >= threeMonthsAgo && sykmelding.status === gamleSMStatuser.SENDT;
-    });
-};
-
-const unikeArbeidsgivereMedSykmeldingSiste3Maneder = (sykmeldinger: Sykmelding[]) => {
-    const sykmeldingerSiste3Maneder = aktiveSykmeldingerFraSiste3Maneder(sykmeldinger)
-
-    const arbeidsgiverFromSykmeldinger = sykmeldingerToArbeidsgiver(sykmeldingerSiste3Maneder);
-
-    return uniqueArbeidsgivere(arbeidsgiverFromSykmeldinger);
-}
 
 const Pengestopp = ({ brukernavn, sykmeldinger, flaggperson }: IPengestoppProps) => {
     const [modalIsOpen, setModelIsOpen] = useState(false);
