+++ src/reducers/flaggperson.ts
@@ -1,4 +1,11 @@
-import { ENDRE_STATUS_FEILET, ENDRER_STATUS, HENT_STATUS_FEILET, HENTER_STATUS, STATUS_ENDRET, STATUS_HENTET } from '../actions/flaggperson_actions';
+import {
+    ENDRE_STATUS_FEILET,
+    ENDRER_STATUS,
+    HENT_STATUS_FEILET,
+    HENTER_STATUS,
+    STATUS_ENDRET,
+    STATUS_HENTET
+} from '../actions/flaggperson_actions';
 import { StatusEndring } from '../types/FlaggPerson';
 
 export const initialState = {
